+++ src/app/api/v1/users/[id]/route.ts
@@ -4,6 +4,10 @@ import {NextRequest, NextResponse} from 'next/server'
 import prisma from '@/lib/db'
 
 export async function GET(request: NextRequest, { params }: { params: { id: string } }) {
+
+    if (request.method !== 'GET'){
+        return NextResponse.json({error: "Method not allowed"}, {status : 405})
+    }
     try {
         const user = await prisma.user.findUnique({
             where: { id: params.id },
@@ -27,7 +31,7 @@ export async function GET(request: NextRequest, { params }: { params: { id: stri
         // user.firstName = decrypt(user.firstName)
         // user.lastName = decrypt(user.lastName)
 
-        return NextResponse.json(user)
+        return NextResponse.json(user, {status : 200})
     } catch (error) {
         console.error('Error fetching user:', error)
         return NextResponse.json({ error: 'Failed to fetch user' }, { status: 500 })
@@ -35,6 +39,10 @@ export async function GET(request: NextRequest, { params }: { params: { id: stri
 }
 
 export async function PUT(request: NextRequest, { params }: { params: { id: string } }) {
+    if (request.method !== 'PUT'){
+        return NextResponse.json({error: "Method not allowed"}, {status : 405})
+    }
+
     try {
         const body = await request.json()
         const { firstName, lastName, email, role } = body
@@ -65,20 +73,24 @@ export async function PUT(request: NextRequest, { params }: { params: { id: stri
         // updatedUser.firstName = decrypt(updatedUser.firstName)
         // updatedUser.lastName = decrypt(updatedUser.lastName)
 
-        return NextResponse.json(updatedUser)
+        return NextResponse.json(updatedUser, {status : 201})
     } catch (error) {
         console.error('Error updating user:', error)
         return NextResponse.json({ error: 'Failed to update user' }, { status: 500 })
     }
 }
 
-export async function DELETE(request: Request, { params }: { params: { id: string } }) {
+export async function DELETE(request: NextRequest, { params }: { params: { id: string } }) {
+
+    if (request.method !== 'DELETE'){
+        return NextResponse.json({error: "Method not allowed"}, {status : 405})
+    }
     try {
         await prisma.user.delete({
             where: { id: params.id },
         })
 
-        return NextResponse.json({ message: 'User deleted successfully' })
+        return NextResponse.json({ message: 'User deleted successfully' }, {status : 200})
     } catch (error) {
         console.error('Error deleting user:', error)
         return NextResponse.json({ error: 'Failed to delete user' }, { status: 500 })
